+++ iOSReferenceApp/EnvironmentSelection/ViewModels/EnvironmentSelectionViewModel.swift
@@ -22,7 +22,7 @@ class EnvironmentSelectionViewModel {
         return environments[index].customers.map{ $0.pickerModelTitle }
     }
     
-    var updatedDefaultValues: (_ baseUrl: String?, _ customer: String?, _ businessUnit: String?) -> Void = { _, _, _ in }
+    var updatedDefaultValues: (_ baseUrl: String?, _ customer: String?, _ businessUnit: String?, _ usesMfa: Bool?) -> Void = { _,_,_,_ in }
     var updatedPresets: (_ environment: String, _ customer: String, _ enableCustomer: Bool) -> Void = { _, _, _ in }
     
     init() {
@@ -35,7 +35,7 @@ class EnvironmentSelectionViewModel {
         selectedCustomer = nil
         
         updatedPresets("Environment", "Customer", false)
-        updatedDefaultValues(nil, nil, nil)
+        updatedDefaultValues(nil, nil, nil, nil)
     }
     
     func select(environment: String) {
@@ -44,7 +44,7 @@ class EnvironmentSelectionViewModel {
         selectedCustomer = nil
         
         updatedPresets(environmentSelections[index], "Customer", true)
-        updatedDefaultValues(environments[index].url, nil, nil)
+        updatedDefaultValues(environments[index].url, nil, nil, nil)
         
         if let preselectedCustomer = customerSelections(index: 0).first {
             select(customer: preselectedCustomer)
@@ -61,13 +61,8 @@ class EnvironmentSelectionViewModel {
                        true)
         updatedDefaultValues(environments[index].url,
                              environments[environmentIndex].customers[index].customer,
-                             environments[environmentIndex].customers[index].businessUnit)
-    }
-    
-    var preferedLoginMethod: CustomerConfig.PresetMethod {
-        guard let environmentIndex = selectedEnvironment, let customerIndex = selectedCustomer else { return .login(username: "", password: "", mfa: false) }
-        guard let presetMethod = environments[environmentIndex].customers[customerIndex].presetMethod else { return .login(username: "", password: "", mfa: false) }
-        return presetMethod
+                             environments[environmentIndex].customers[index].businessUnit,
+                             environments[environmentIndex].customers[index].usesMfa)
     }
     
     var selectedExposureEnvironment: Environment? {