+++ iOSReferenceApp/Models/UserInfo/UserInfo.swift
@@ -46,9 +46,8 @@ enum UserInfo {
                            businessUnit: businessUnit)
     }
     
-    static var environmentLoginMethod: CustomerConfig.PresetMethod? {
-        guard let string = TinyDB.getString(UserInfo.Lets.KEY_ENVIRONMENT_LOGIN_METHOD) else { return nil }
-        return CustomerConfig.PresetMethod(persistenceString: string)
+    static var environmentUsesMfa: Bool {
+        return TinyDB.getBool(UserInfo.Lets.KEY_ENVIRONMENT_LOGIN_METHOD)
     }
     
     struct Lets {
@@ -130,7 +129,7 @@ extension UserInfo {
         TinyDB.save(environment.businessUnit, withKey: UserInfo.Lets.KEY_CUSTOMER_BUSINESS_UNIT)
     }
     
-    static func environment(loginMethod: CustomerConfig.PresetMethod) {
-        TinyDB.save(loginMethod.persistenceString, withKey: UserInfo.Lets.KEY_ENVIRONMENT_LOGIN_METHOD)
+    static func environment(usesMfa: Bool) {
+        TinyDB.save(usesMfa, withKey: UserInfo.Lets.KEY_ENVIRONMENT_LOGIN_METHOD)
     }
 }